\name{Ridit-package}
\alias{Ridit-package}
\alias{Ridit}
\docType{package}
\title{
Ridit Analysis (An extension of the Kruskal-Wallis Test.)
}
\description{
An extension of the Kruskal-Wallis Test that allow selection of arbitrary reference group. Also provide Mean Ridit for each group. Mean Ridit of a group is an estimate of probability a random observation from that group will be greater than or equal to a random observation from reference group.
}
\details{
\tabular{ll}{
Package: \tab Ridit\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2012-10-15\cr
License: \tab GPL-2 | GPL-3\cr
}

}
\author{
SeyedMahmood TaghaviShahri
Maintainer: SeyedMahmood TaghaviShahri <taghavi_m@razi.tums.ac.ir>
}
\references{
Fleiss, J. L., (1986), The Design and Analysis of Clinical Experiments. New York: John Wiley & Sons. 
}
\keyword{ package }


